requirements.txt
eventlet==0.36.1
Flask==3.0.3
numpy==1.22.1
opencv-python==4.9.0.80
pandas==2.0.3
pillow==10.2.0
python-socketio==4.2.1
scikit-learn==1.3.2
torch==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.66.2
